feat: move tasks resourses to ts
