--- src/resources/tasks/task.model.js
@@ -1,27 +1,0 @@
-const uuid = require('uuid');
-
-class Task {
-  constructor({
-    id = uuid.v4(),
-    title = 'My Task',
-    order = 0,
-    description = 'Task description',
-    userId = null,
-    boardId = null,
-    columnId = null
-  } = {}) {
-    this.id = id;
-    this.title = title;
-    this.order = order;
-    this.description = description;
-    this.userId = userId;
-    this.boardId = boardId;
-    this.columnId = columnId
-  }
-
-  static toResponse(task) {
-    return task;
-  }
-}
-
-module.exports = Task;
